+++ tests/style/bug_163_trailing_semicolons/test_07_d.m
@@ -1,0 +1,10 @@
+% (C) Copyright 2020 Florian Schanda
+
+d = [
+    ;
+    ;
+
+    1
+    ;
+    ]
+% d = 1